Setup an initial interaction with classes
Added a representation of the classes to represent the elevator
interaction.
An elevator has N buttons for the floors, each floor has one or 2
buttons to call the elevator.
The buttons communicate with the elevator via a EventListener which
sends 3 distinct types of events (so far).
Added a nice terminal interface that should be easy to use and also
display the state of the elevator...
Now the only thing missing is a life cicle